pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "zoo"
version = "0.1.0"
description = "Web visualization for PANDA_CORE"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "pyserial>=3.5",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "httpx>=0.27"]
